requirements.txt
streamlit
streamlit-authenticator
pandas
openpyxl
python-docx
jinja2
pyyaml
bcrypt